refactor: main-button props 구조 변경으로 인한 로직수정
